#52 create query handler for get long lat  data
