+++ repository/repository_interface.go
@@ -15,6 +15,7 @@ type PhoneBookI interface {
 	GetCategoryNameByID(ctx context.Context, id int64) (string, error)
 	GetLocationNameByID(ctx context.Context, id int64) (*string, error)
 	GetListPhonebookByLongLat(ctx context.Context, params *model.GetListRequest) ([]*model.PhoneBookResponse, error)
+	GetListPhonebookByLongLatMeta(ctx context.Context, params *model.GetListRequest) (int64, error)
 	// Create section
 	Insert(ctx context.Context, params *model.AddPhonebook) error
 	// Update section
